+++ src/Manager/Native.php
@@ -26,11 +26,16 @@ final class Native implements Manager
 {
     private ?Session\Id $session = null;
 
-    public function __construct(Path $save = null)
+    private function __construct(Path $save = null)
     {
         if ($save instanceof Path) {
             $_ = \session_save_path($save->toString());
         }
+    }
+
+    public static function of(Path $save = null): self
+    {
+        return new self($save);
     }
 
     public function start(ServerRequest $request): Maybe